fix: allow lagging/failed Sync Health in Platform Store
Expand base_datasets_sync_health_check so Incremental progress can persist
lagging/ok labels. Treat cutover checkpoint alone as unknown Sync Health
until Incremental applies.

+++ crates/app/tests/deployment_runtime_observability.rs
@@ -33,7 +33,7 @@ fn mongo_host() -> String {
     std::env::var("MIGRALOOP_TEST_MONGO_HOST").unwrap_or_else(|_| "127.0.0.1".to_string())
 }
 
-fn mongo_port() -> u16 {
+fn mongo_port() -> i32 {
     std::env::var("MIGRALOOP_TEST_MONGO_PORT")
         .ok()
         .and_then(|v| v.parse().ok())
